requirements.txt
flask>=3.0.0
gunicorn>=22.0.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
reportlab>=4.0.0
python-docx>=1.1.0
